refactor: abstract VideoOverlay and SearchInput components and fix rerenderig
