+++ frontend/src/components/SearchInput/SearchInput.tsx
