+++ frontend/src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './InfiniteGallery'
 export * from './PageContainer'
+export * from './VideoOverlay'
+export * from './SearchInput'
